fix: Eliminate all compiler sign-compare warnings
- Added (size_t) casts to capacity comparisons with SIZE_MAX
- Fixed 5 warnings in transpiler.c:
  * sb_append StringBuilder capacity check (line 32)
  * add_module_header capacity check (line 163)
  * register_tuple_type capacity check (line 367)
  * register_function_signature capacity check (line 436)
- Fixed 1 warning in transpiler_iterative_v3_twopass.c:
  * worklist_grow capacity check (line 77)

All comparisons now properly cast int to size_t before
comparing with SIZE_MAX, eliminating sign mismatch warnings.

Build is now fully green with zero warnings ✅

Test results:
- make clean bootstrap test examples: 0 warnings/errors
- All 102+ tests pass
- All 27 compiled examples build successfully

Related to nanolang-929.

+++ src/transpiler.c
@@ -29,7 +29,7 @@ void sb_append(StringBuilder *sb, const char *str) {
     int len = safe_strlen(str);
     while (sb->length + len >= sb->capacity) {
         /* Check for overflow before doubling capacity */
-        if (sb->capacity > SIZE_MAX / 2) {
+        if ((size_t)sb->capacity > SIZE_MAX / 2) {
             fprintf(stderr, "Error: StringBuilder capacity overflow\n");
             exit(1);
         }
@@ -160,7 +160,7 @@ static void add_module_header(const char *header, int priority) {
     /* Expand capacity if needed */
     if (g_module_headers_count >= g_module_headers_capacity) {
         int new_capacity = g_module_headers_capacity == 0 ? 8 : g_module_headers_capacity * 2;
-        if (new_capacity > SIZE_MAX / sizeof(ModuleHeader)) {
+        if ((size_t)new_capacity > SIZE_MAX / sizeof(ModuleHeader)) {
             fprintf(stderr, "Error: Module headers capacity overflow\n");
             exit(1);
         }
@@ -364,7 +364,7 @@ static const char *register_tuple_type(TupleTypeRegistry *reg, TypeInfo *info) {
     
     /* Register new tuple type */
     if (reg->count >= reg->capacity) {
-        if (reg->capacity > SIZE_MAX / 2) {
+        if ((size_t)reg->capacity > SIZE_MAX / 2) {
             fprintf(stderr, "Error: Tuple registry capacity overflow\n");
             exit(1);
         }
@@ -433,7 +433,7 @@ static const char *register_function_signature(FunctionTypeRegistry *reg, Functi
     
     /* Register new signature */
     if (reg->count >= reg->capacity) {
-        if (reg->capacity > SIZE_MAX / 2) {
+        if ((size_t)reg->capacity > SIZE_MAX / 2) {
             fprintf(stderr, "Error: Function registry capacity overflow\n");
             exit(1);
         }
